requirements.txt
Flask
Flask-WTF
Flask-SQLAlchemy
Flask-Limiter
python-dotenv
requests